fix: preview toolbar tooltips staying open
If Tooltips or Popovers are attached to a button which can become
disabled, then the tooltip can become stuck visible or hidden.

This is because disabled buttons do not emit pointer events. This
problem is described in the Blueprint documentation:

https://blueprintjs.com/docs/#core/components/tooltip.props

The solution is to replace the Button component with AnchorButton.
It appears identical but maintains correct behavior when disabled.