fix: menus not dismissing from canvas clicks
The Babylon scene by default sets its preventDefaultOnPointerDown
value to true. This causes Blueprint to not observe the mousedown
event it is looking for in order to dismiss an open Popover, like
the ones used for the main toolbar.

This could be solved by setting preventDefaultOnPointerDown to
false, but this could have unanticipated consequences.

Instead, I addressed the issue by adding the `hasBackdrop` prop to
the Blueprint Popover components. This is documented here:

https://blueprintjs.com/docs/#core/components/popover.backdrop

Essentially, this prop adds a click capturing overlay to the app
while the popover is open, causing a click anywhere on the app to
close it -- including the canvas.